refactor: split dashboard/_api_health.py into health-check and spawn modules
Extract health polling into _api_health_check.py and process spawning
into _api_spawn.py. Original _api_health.py remains as re-export entry point.
